+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Harmonised FlowAPI parameter names for start and end dates. They are now all `start_date` and `end_date`
 - Further improvements to token display in FlowAuth. [#1124](https://github.com/Flowminder/FlowKit/issues/1124)
 - Increased the FlowDB quickstart container's timeout to 15 minutes. [#782](https://github.com/Flowminder/FlowKit/issues/782)
+- `Union` and `Query.union` now accept a variable number of queries to concatenate. [#4565](https://github.com/Flowminder/FlowKit/issues/4565)
 
 ### Fixed
 - Autoflow's prefect version is now current. [#2544](https://github.com/Flowminder/FlowKit/issues/2544)